feat : geojson api (info, geojson 데이터) 완료
추가 요구사항 완료

filterList : int 에서 List로 변경
Dto에서 데이터 가공 후 서비스 코드로 전달 완료

+++ src/main/java/com/hipspot/web/dto/GeometryDto.java
@@ -8,9 +8,9 @@ import java.util.List;
 public class GeometryDto {
 
     private String type;
-    private List<Long> coordinates;
+    private List<Double> coordinates;
 
-    public GeometryDto(String type, List<Long> coordinates) {
+    public GeometryDto(String type, List<Double> coordinates) {
         this.type = type;
         this.coordinates = coordinates;
     }
